app/requirements.txt
Flask==0.12.4
flask-restplus
flask-marshmallow
ipython
